requirements.txt
absl-py==0.12.0
appdirs==1.4.4
attrs==20.3.0
audioread==2.1.9
Bottleneck==1.3.2
brotlipy==0.7.0
cached-property==1.5.2
cachetools==4.2.1
certifi==2021.5.30
cffi @ file:///tmp/build/80754af9/cffi_1613246945912/work
chainer==6.0.0
chardet @ file:///tmp/build/80754af9/chardet_1607706746162/work
click==7.1.2
colorama==0.4.4
conda==4.10.1
conda-package-handling @ file:///tmp/build/80754af9/conda-package-handling_1618262148928/work
ConfigArgParse==1.4
configparser==5.0.2
cryptography @ file:///tmp/build/80754af9/cryptography_1616769286105/work
ctc-segmentation==1.5.2
cycler==0.10.0
Cython==0.29.22
dataclasses==0.6
decorator==5.0.5
Distance==0.1.3
dlp-mpi==0.0.3
docker-pycreds==0.4.0
docopt==0.6.2
editdistance==0.5.2
einops==0.3.0
-e git+https://github.com/espnet/espnet@c2f56b113ad65b045f5bbc332fd7a3581c8304d4#egg=espnet
espnet-model-zoo==0.0.0a28
espnet-tts-frontend==0.0.3
fairscale==0.3.3
fastdtw==0.3.4
ffmpeg-python==0.2.0
filelock==3.0.12
fire==0.4.0
flake8==3.9.2
future==0.18.2
g2p-en==2.1.0
gdown==3.12.2
gitdb==4.0.7
GitPython==3.1.14
google-auth==1.28.0
google-auth-oauthlib==0.4.4
grpcio==1.36.1
h5py==3.2.1
humanfriendly==9.1
idna @ file:///home/linux1/recipes/ci/idna_1610986105248/work
inflect==5.3.0
jaconv==0.3
joblib @ file:///tmp/build/80754af9/joblib_1613502643832/work
jsonpickle==1.5.2
jsonschema==3.2.0
kaldiio==2.17.2
kiwisolver==1.3.1
lazy-dataset==0.0.11
librosa==0.8.0
llvmlite==0.36.0
Markdown==3.3.4
matplotlib==3.1.0
mccabe==0.6.1
mir-eval==0.6
mkl-fft==1.3.0
mkl-random @ file:///tmp/build/80754af9/mkl_random_1618853849286/work
mkl-service==2.3.0
Morfessor==2.0.6
mpi4py==3.0.3
munch==2.5.0
musdb==0.4.0
museval==0.4.0
nara-wpe==0.0.7
nltk==3.5
nnmnkwii==0.0.22
numba==0.53.1
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1620830962040/work
oauthlib==3.1.0
packaging==20.9
paderbox @ git+http://github.com/fgnt/paderbox@2e3a944b70d1128f6c75b7205c46b0d8fa1dd69a
pandas==1.2.3
pathtools==0.1.2
-e git+https://github.com/fgnt/pb_bss.git@96fd72cb5934fb3ec21a707cc54ac6263782a71a#egg=pb_bss&subdirectory=../../pb_bss
-e git+https://github.com/fgnt/pb_chime5.git@16f36f90be93b568c647e4a9f0e15e153a16c2a8#egg=pb_chime5
Pillow==8.2.0
pooch==1.3.0
promise==2.3
protobuf==3.7.1
psutil==5.8.0
py-cpuinfo==8.0.0
pyaml==20.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.6.2
pycodestyle==2.7.0
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pyflakes==2.3.1
pymongo==3.11.4
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyparsing==2.4.7
pypinyin==0.41.0
pyrsistent==0.17.3
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
pysptk==0.1.18
pystoi==0.3.3
python-dateutil==2.8.1
pytorch-ranger==0.1.1
pytorch-wpe==0.0.1
pytz==2021.1
pyworld==0.2.12
PyYAML==5.4.1
regex==2021.4.4
requests @ file:///tmp/build/80754af9/requests_1608241421344/work
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.7.2
ruamel-yaml-conda @ file:///tmp/build/80754af9/ruamel_yaml_1616016699510/work
sacred==0.8.2
scikit-learn @ file:///tmp/build/80754af9/scikit-learn_1621370412049/work
scipy @ file:///tmp/build/80754af9/scipy_1618855647378/work
sentencepiece==0.1.86
sentry-sdk==1.0.0
shortuuid==1.0.1
simplejson==3.17.2
six @ file:///tmp/build/80754af9/six_1623709665295/work
smmap==4.0.0
SoundFile==0.10.3.post1
sox==1.4.1
stempeg==0.2.3
subprocess32==3.5.4
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.2
termcolor==1.1.0
threadpoolctl @ file:///tmp/tmp9twdgx9k/threadpoolctl-2.1.0-py3-none-any.whl
torch==1.8.0
torch-complex==0.2.1
torch-optimizer==0.1.0
torchaudio==0.8.0
tqdm @ file:///tmp/build/80754af9/tqdm_1615925068909/work
typeguard==2.12.0
typing-extensions @ file:///home/ktietz/src/ci_mi/typing_extensions_1612808209620/work
Unidecode==1.2.0
urllib3 @ file:///tmp/build/80754af9/urllib3_1615837158687/work
wandb==0.10.25
-e git+https://github.com/HawkAaron/warp-transducer.git@f546575109111c455354861a0567c8aa794208a2#egg=warprnnt_pytorch&subdirectory=pytorch_binding
Werkzeug==1.0.1
wrapt==1.12.1
yamlargparse==1.31.1
youtube-dl==2021.4.1
